add entity attached Gui

+++ src/main/java/lib/plenilune/client/render/gui/GuiRender.java
@@ -37,15 +37,15 @@ public class GuiRender {
 	 * @param minV
 	 * @param maxV
 	 */
-	public static void blitImage(PoseStack poseStack, net.minecraft.resources.ResourceLocation atlasLocation, int x1, int x2, int y1, int y2, int blitOffset, float minU, float maxU, float minV, float maxV) {
+	public static void blitImage(PoseStack poseStack, net.minecraft.resources.ResourceLocation atlasLocation, float x1, float x2, float y1, float y2, float blitOffset, float minU, float maxU, float minV, float maxV) {
 		RenderSystem.setShaderTexture(0, (atlasLocation));
 		RenderSystem.setShader(GameRenderer::getPositionTexShader);
 		Matrix4f matrix4f = poseStack.last().pose();
 		BufferBuilder bufferBuilder = TesselatorInstance.getBufferBuilder(VertexFormat.Mode.QUADS, DefaultVertexFormat.POSITION_TEX);
-		TesselatorInstance.posUvVertex(bufferBuilder, matrix4f, (float) x1, (float) y1, (float) blitOffset, minU, minV);
-		TesselatorInstance.posUvVertex(bufferBuilder, matrix4f, (float) x1, (float) y2, (float) blitOffset, minU, maxV);
-		TesselatorInstance.posUvVertex(bufferBuilder, matrix4f, (float) x2, (float) y2, (float) blitOffset, maxU, maxV);
-		TesselatorInstance.posUvVertex(bufferBuilder, matrix4f, (float) x2, (float) y1, (float) blitOffset, maxU, minV);
+		TesselatorInstance.posUvVertex(bufferBuilder, matrix4f, x1, y1, blitOffset, minU, minV);
+		TesselatorInstance.posUvVertex(bufferBuilder, matrix4f, x1, y2, blitOffset, minU, maxV);
+		TesselatorInstance.posUvVertex(bufferBuilder, matrix4f, x2, y2, blitOffset, maxU, maxV);
+		TesselatorInstance.posUvVertex(bufferBuilder, matrix4f, x2, y1, blitOffset, maxU, minV);
 		TesselatorInstance.drawBufferBuilder(bufferBuilder);
 	}
